backend/requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
fernet==1.0.1
Flask==3.1.3
flask-cors==6.0.2
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
playwright==1.58.0
psycopg2-binary==2.9.11
pyaes==1.6.1
pycparser==3.0
pyee==13.0.1
Pygments==2.19.2
python-dotenv==1.2.2
requests==2.32.5
rich==14.3.3
somtoday==0.1.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
